rename to SearchStatisticsHooks
